+++ spec/rubocop/target_finder_spec.rb
@@ -97,7 +97,7 @@ RSpec.describe RuboCop::TargetFinder, :isolated_environment do
       let(:args) { [] }
 
       it 'finds files under the current directory' do
-        Dir.chdir('dir1') do
+        RuboCop::PathUtil.chdir('dir1') do
           expect(found_files.empty?).to be(false)
           found_files.each do |file|
             expect(file).to include('/dir1/')
@@ -111,7 +111,7 @@ RSpec.describe RuboCop::TargetFinder, :isolated_environment do
       let(:args) { ['../dir2'] }
 
       it 'finds files under the specified directory' do
-        Dir.chdir('dir1') do
+        RuboCop::PathUtil.chdir('dir1') do
           expect(found_files.empty?).to be(false)
           found_files.each do |file|
             expect(file).to include('/dir2/')